Remove disconnect link from typeracer

+++ app/decorators/provider_decorator.rb
@@ -18,7 +18,7 @@ class ProviderDecorator < Draper::Decorator
   end
 
   def delete_link
-    if connected?
+    if oauthable? && connected?
       h.link_to "Disconnect", h.provider_destroy_path(self), method: :delete, "data-confirm": "Are you sure?"
     else
       "-"
